Simplify CONTEXT API
`candidates` now returns either `Ok opam` or `Error pkg` for each
package. This is clearer than using an option type and avoids the need
for a separate `load` function. It also makes it possible to filter
packages based on the content of the opam file without having to load it
twice.

We also no longer bother loading the opam file for rejects. All we need
is the name.